Move parameter definitions into sub-directories
