+++ benchmark/_operations.py
@@ -3,11 +3,7 @@
 Operations for the control service benchmarks.
 """
 
-from pyrsistent import PClass, field
-from zope.interface import Interface, implementer
-
-from twisted.internet.defer import succeed
-from twisted.web.client import ResponseFailed
+from zope.interface import Interface
 
 
 class IProbe(Interface):
@@ -39,64 +35,3 @@ class IOperation(Interface):
         perform real work, the operation may return a different
         probe each time.
         """
-
-
-def _report_error(failure):
-    failure.trap(ResponseFailed)
-    for reason in failure.value.reasons:
-        reason.printTraceback()
-    return reason
-
-
-@implementer(IProbe)
-class _NoOpRequest(PClass):
-    """
-    A probe that performs no operation.
-    """
-
-    def run(self):
-        return succeed(None)
-
-    def cleanup(self):
-        return succeed(None)
-
-
-@implementer(IOperation)
-class _NoOperation(PClass):
-    """
-    An nop operation.
-    """
-
-    control_service = field(mandatory=True)
-
-    def get_probe(self):
-        return _NoOpRequest()
-
-
-@implementer(IProbe)
-class _ReadRequest(PClass):
-    """
-    A probe to perform a read request on the control service.
-    """
-
-    control_service = field(mandatory=True)
-
-    def run(self):
-        d = self.control_service.list_datasets_state()
-        d.addErrback(_report_error)
-        return d
-
-    def cleanup(self):
-        return succeed(None)
-
-
-@implementer(IOperation)
-class _ReadRequestOperation(PClass):
-    """
-    An operation to perform a read request on the control service.
-    """
-
-    control_service = field(mandatory=True)
-
-    def get_probe(self):
-        return _ReadRequest(control_service=self.control_service)
