+++ benchmark/metrics/__init__.py
@@ -1,0 +1,1 @@
+from .wallclock import WallClock
