+++ benchmark/operations/__init__.py
@@ -1,0 +1,2 @@
+from .no_op import NoOperation
+from .read_request import ReadRequest
