+++ benchmark/scenarios/__init__.py
@@ -1,0 +1,1 @@
+from .no_load import NoLoadScenario
